Shifting decorator tests to new file
